A lot
The biggest thing is that i added the Project settings gui, so everything is prepared for the implementation into the AutoComplete of addons.

I also added a Auto saver that loops every 25 seconds, further the code is saved on a Focus loss of the code area.

I noticed the highlighting was throwing a Stackoverflow error when adding quotes at a specific position in a skript.
Wíll have to Check this later